chore: sync local version with NPM 0.9.3-59
- Update version from 0.9.2-59 to 0.9.3-59
- Fix setAudioGuidanceType parameter type from int to double to handle React Native bridge values
- Improve showRouteOverview implementation with proper UI thread handling and error catching
- Add missing UiThreadUtil import in NavViewFragment
- Update changelog with version 0.9.3 changes

+++ android/src/main/java/com/google/android/react/navsdk/NavModule.java
@@ -609,14 +609,14 @@ public class NavModule extends ReactContextBaseJavaModule
   }
 
   @ReactMethod
-  public void setAudioGuidanceType(int jsValue) {
+  public void setAudioGuidanceType(double jsValue) {
     if (mNavigator == null) {
       return;
     }
 
     UiThreadUtil.runOnUiThread(
         () -> {
-          mNavigator.setAudioGuidance(EnumTranslationUtil.getAudioGuidanceFromJsValue(jsValue));
+          mNavigator.setAudioGuidance(EnumTranslationUtil.getAudioGuidanceFromJsValue((int) jsValue));
         });
   }
